Rename test files to ensure dockerTag runs
